+++ app/frontend/api/productApiHandler.js
@@ -21,3 +21,21 @@ export async function fetchProducts(token) {
     }
     return res.json()
 }
+
+/**
+ * Fetch products information from backend
+ * @param {string} token
+ * @return {Promise<Object>}
+ * @throws {ApiError}
+ */
+
+export async function fetchProductById(productId) {
+    const res = await fetch(`${config.host}/products/${productId}`);
+
+    if (!res.ok) {
+        const errorData = await res.json()
+        const errorMessage = errorData.message || "Failed to fetch product"
+        throw new ApiError(errorMessage, res.status, errorData)
+    }
+    return await res.json();
+}
