+++ app/frontend/pages/products/productDetailled/ProductDetailled.js
@@ -1,11 +1,17 @@
-import { fetchProductById } from "../../../api/productDetailledApiHandler.js"
+import { fetchProductById } from "../../../api/productApiHandler.js"
 
-export default async function ProductDetailledLoader({ params }) {
-  const productId = params.productId;
+export default async function ProductDetailledLoader(req, res) {
+  const productId = req.params.productId;
   const product = await fetchProductById(productId);
+
+  //proof if admin is connected 
+  const isAdmin = req?.user?.isAdmin
+console.log(req.user)
+
 
   return {
     title: product.name,
-    product: product
+    product: product,
+    isAdmin: isAdmin
   };
 }
